# admin-service/api/api.gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for BookingStatus.
const (
	BookingStatusConfirmed BookingStatus = "confirmed"
	BookingStatusPending   BookingStatus = "pending"
	BookingStatusRejected  BookingStatus = "rejected"
)

// Defines values for GetApiBookingsParamsStatus.
const (
	GetApiBookingsParamsStatusConfirmed GetApiBookingsParamsStatus = "confirmed"
	GetApiBookingsParamsStatusPending   GetApiBookingsParamsStatus = "pending"
	GetApiBookingsParamsStatusRejected  GetApiBookingsParamsStatus = "rejected"
)

// Booking defines model for Booking.
type Booking struct {
	BookingId          *string              `json:"booking_id,omitempty"`
	Checkin            *openapi_types.Date  `json:"checkin,omitempty"`
	Checkout           *openapi_types.Date  `json:"checkout,omitempty"`
	ConfirmationNumber *string              `json:"confirmation_number"`
	GuestEmail         *openapi_types.Email `json:"guest_email,omitempty"`
	GuestName          *string              `json:"guest_name,omitempty"`
	Guests             *int                 `json:"guests,omitempty"`
	HotelId            *string              `json:"hotel_id,omitempty"`
	Status             *BookingStatus       `json:"status,omitempty"`
	TotalPrice         *float32             `json:"total_price,omitempty"`
}

// BookingStatus defines model for Booking.Status.
type BookingStatus string

// Error defines model for Error.
type Error struct {
	Error *string `json:"error,omitempty"`
}

// GetApiBookingsParams defines parameters for GetApiBookings.
type GetApiBookingsParams struct {
	// Status Filter by booking status
	Status *GetApiBookingsParamsStatus `form:"status,omitempty" json:"status,omitempty"`
}

// GetApiBookingsParamsStatus defines parameters for GetApiBookings.
type GetApiBookingsParamsStatus string

// PostApiBookingsBookingIdRejectJSONBody defines parameters for PostApiBookingsBookingIdReject.
type PostApiBookingsBookingIdRejectJSONBody struct {
	// Reason Reason for rejection
	Reason string `json:"reason"`
}

// PostApiBookingsBookingIdRejectJSONRequestBody defines body for PostApiBookingsBookingIdReject for application/json ContentType.
type PostApiBookingsBookingIdRejectJSONRequestBody PostApiBookingsBookingIdRejectJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get bookings
	// (GET /api/bookings)
	GetApiBookings(w http.ResponseWriter, r *http.Request, params GetApiBookingsParams)
	// Get booking by ID
	// (GET /api/bookings/{booking_id})
	GetApiBookingsBookingId(w http.ResponseWriter, r *http.Request, bookingId string)
	// Approve booking
	// (POST /api/bookings/{booking_id}/approve)
	PostApiBookingsBookingIdApprove(w http.ResponseWriter, r *http.Request, bookingId string)
	// Reject booking
	// (POST /api/bookings/{booking_id}/reject)
	PostApiBookingsBookingIdReject(w http.ResponseWriter, r *http.Request, bookingId string)
	// Get flag status
	// (GET /api/flags)
	GetApiFlags(w http.ResponseWriter, r *http.Request)
	// Health check
	// (GET /health)
	GetHealth(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetApiBookings operation middleware
func (siw *ServerInterfaceWrapper) GetApiBookings(w http.ResponseWriter, r *http.Request) {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetApiBookingsParams

	// ------------- Optional query parameter "status" -------------

	err = runtime.BindQueryParameter("form", true, false, "status", r.URL.Query(), &params.Status)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "status", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetApiBookings(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetApiBookingsBookingId operation middleware
func (siw *ServerInterfaceWrapper) GetApiBookingsBookingId(w http.ResponseWriter, r *http.Request) {
	var err error

	// ------------- Path parameter "booking_id" -------------
	var bookingId string

	err = runtime.BindStyledParameterWithOptions("simple", "booking_id", r.PathValue("booking_id"), &bookingId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "booking_id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetApiBookingsBookingId(w, r, bookingId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostApiBookingsBookingIdApprove operation middleware
func (siw *ServerInterfaceWrapper) PostApiBookingsBookingIdApprove(w http.ResponseWriter, r *http.Request) {
	var err error

	// ------------- Path parameter "booking_id" -------------
	var bookingId string

	err = runtime.BindStyledParameterWithOptions("simple", "booking_id", r.PathValue("booking_id"), &bookingId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "booking_id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostApiBookingsBookingIdApprove(w, r, bookingId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostApiBookingsBookingIdReject operation middleware
func (siw *ServerInterfaceWrapper) PostApiBookingsBookingIdReject(w http.ResponseWriter, r *http.Request) {
	var err error

	// ------------- Path parameter "booking_id" -------------
	var bookingId string

	err = runtime.BindStyledParameterWithOptions("simple", "booking_id", r.PathValue("booking_id"), &bookingId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "booking_id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostApiBookingsBookingIdReject(w, r, bookingId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetApiFlags operation middleware
func (siw *ServerInterfaceWrapper) GetApiFlags(w http.ResponseWriter, r *http.Request) {
	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetApiFlags(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetHealth operation middleware
func (siw *ServerInterfaceWrapper) GetHealth(w http.ResponseWriter, r *http.Request) {
	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("GET "+options.BaseURL+"/api/bookings", wrapper.GetApiBookings)
	m.HandleFunc("GET "+options.BaseURL+"/api/bookings/{booking_id}", wrapper.GetApiBookingsBookingId)
	m.HandleFunc("POST "+options.BaseURL+"/api/bookings/{booking_id}/approve", wrapper.PostApiBookingsBookingIdApprove)
	m.HandleFunc("POST "+options.BaseURL+"/api/bookings/{booking_id}/reject", wrapper.PostApiBookingsBookingIdReject)
	m.HandleFunc("GET "+options.BaseURL+"/api/flags", wrapper.GetApiFlags)
	m.HandleFunc("GET "+options.BaseURL+"/health", wrapper.GetHealth)

	return m
}