# admin-service/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Admin Service API",
    "description": "Admin service for managing hotel bookings with Flipt feature flag evaluation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8001",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "description": "Check if the service is healthy",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "healthy"
                    },
                    "service": {
                      "type": "string",
                      "example": "admin-service"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings": {
      "get": {
        "summary": "Get bookings",
        "description": "Retrieve all bookings, optionally filtered by status",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by booking status",
            "schema": {
              "type": "string",
              "enum": ["pending", "confirmed", "rejected"],
              "default": "pending"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of bookings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bookings": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Booking"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/{booking_id}": {
      "get": {
        "summary": "Get booking by ID",
        "description": "Retrieve a specific booking by its ID",
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "description": "The booking ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Booking"
                }
              }
            }
          },
          "404": {
            "description": "Booking not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/{booking_id}/approve": {
      "post": {
        "summary": "Approve booking",
        "description": "Approve a pending booking using Flipt feature flags to determine auto-approval and tier",
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "description": "The booking ID to approve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking approved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "booking_id": {
                      "type": "string"
                    },
                    "auto_approved": {
                      "type": "boolean"
                    },
                    "approval_tier": {
                      "type": "string",
                      "example": "manager"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g., booking already processed)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Booking not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/{booking_id}/reject": {
      "post": {
        "summary": "Reject booking",
        "description": "Reject a pending booking with a reason",
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "description": "The booking ID to reject",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "description": "Reason for rejection"
                  }
                },
                "required": ["reason"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking rejected",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "booking_id": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "reason": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Booking not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/flags": {
      "get": {
        "summary": "Get flag status",
        "description": "Get current status of Flipt feature flags",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Flag status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "auto_approval": {
                      "type": "object",
                      "properties": {
                        "enabled": {
                          "type": "boolean"
                        },
                        "reason": {
                          "type": "string"
                        }
                      }
                    },
                    "approval_tier": {
                      "type": "object",
                      "properties": {
                        "variant": {
                          "type": "string"
                        },
                        "reason": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Booking": {
        "type": "object",
        "properties": {
          "booking_id": {
            "type": "string",
            "example": "BK-12345678"
          },
          "hotel_id": {
            "type": "string",
            "example": "hotel_1"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "confirmed", "rejected"],
            "example": "pending"
          },
          "confirmation_number": {
            "type": "string",
            "nullable": true,
            "example": "CONF-ABC123"
          },
          "total_price": {
            "type": "number",
            "format": "float",
            "example": 450.0
          },
          "guest_name": {
            "type": "string",
            "example": "John Doe"
          },
          "guest_email": {
            "type": "string",
            "format": "email",
            "example": "john@example.com"
          },
          "checkin": {
            "type": "string",
            "format": "date",
            "example": "2025-01-15"
          },
          "checkout": {
            "type": "string",
            "format": "date",
            "example": "2025-01-18"
          },
          "guests": {
            "type": "integer",
            "example": 2
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  }
}